Mejoras formAdmin

+++ ProyectoFinalV1/FormAdmin.cs
@@ -18,9 +18,12 @@ namespace ProyectoFinalV1
     public partial class FormAdmin : Form
     {
 
-        // Variable global para cargar el nombre de las imagenes a cargar
+        // Variable para cargar el nombre de las imagenes a cargar
         string NombreModificar;
         string NombreSubir;
+
+        // Variable para guardar el numero de registros 
+        int registros;
 
         public FormAdmin()
         {
@@ -112,29 +115,41 @@ namespace ProyectoFinalV1
         // Boton para eliminar un producto
         private void button_EliminarProducto_Click(object sender, EventArgs e)
         {
-            try
+
+            // Funcion para contar el numero de registros que hay actualmente
+            registros = Contar_Registros();
+
+            // Checamos si podemos borrar el registro (ya que minimo es 6)
+            if (registros <= 6)
             {
-
-                // Creamos nuestra variable para la base de datos, y pasamos nuestra informacion
-                MySqlConnection conexion = new MySqlConnection("Server=localhost; Database=proyecto; User=root; Password=; Sslmode=none;");
-                // Abrimos nuestra base de datos
-                conexion.Open();
-
-                // Linea de comando de SQL
-                string consulta = "DELETE FROM juegos WHERE ID='" + Convert.ToInt32(textBox_IDBorrar.Text) + "'";
-                // Cargamos nuestro comando
-                MySqlCommand comando = new MySqlCommand(consulta, conexion);
-
-                // Ejecutamos el comando
-                comando.ExecuteNonQuery();
-
-                MessageBox.Show("Producto eliminado correctamente");
-
-                conexion.Close();
+                MessageBox.Show("No es posible borrar el registro (Minimo 6 registros)");
             }
-            catch (Exception ex)
+            else
             {
-                MessageBox.Show("Error al eliminar producto: " + ex.Message);
+                try
+                {
+
+                    // Creamos nuestra variable para la base de datos, y pasamos nuestra informacion
+                    MySqlConnection conexion = new MySqlConnection("Server=localhost; Database=proyecto; User=root; Password=; Sslmode=none;");
+                    // Abrimos nuestra base de datos
+                    conexion.Open();
+
+                    // Linea de comando de SQL
+                    string consulta = "DELETE FROM juegos WHERE ID='" + Convert.ToInt32(textBox_IDBorrar.Text) + "'";
+                    // Cargamos nuestro comando
+                    MySqlCommand comando = new MySqlCommand(consulta, conexion);
+
+                    // Ejecutamos el comando
+                    comando.ExecuteNonQuery();
+
+                    MessageBox.Show("Producto eliminado correctamente");
+
+                    conexion.Close();
+                }
+                catch (Exception ex)
+                {
+                    MessageBox.Show("Error al eliminar el registro: " + ex.Message);
+                }
             }
 
             // Vaciamos el textBox
@@ -145,39 +160,51 @@ namespace ProyectoFinalV1
         // Boton para guardar el nuevo producto
         private void button_Subir_Click(object sender, EventArgs e)
         {
-            try
+
+            // Funcion para contar el numero de registros que hay actualmente
+            registros = Contar_Registros();
+
+            // Checamos si podemos agregar otro registro (ya que el maximo es 10)
+            if (registros >= 10)
             {
-
-                // Creamos nuestra variable para la base de datos, y pasamos nuestra informacion
-                MySqlConnection conexion = new MySqlConnection("Server=localhost; Database=proyecto; User=root; Password=; Sslmode=none;");
-                // Abrimos nuestra base de datos
-                conexion.Open();
-
-                // Linea de comando de SQL
-                string consulta = "INSERT INTO juegos (ID, Nombre, Imagen, Genero, Plataforma, Modalidad, Precio, Stock) VALUES ("
-                    + "'" + textBox_IDSubir.Text + "',"
-                    + "'" + textBox_NombreSubir.Text + "',"
-                    + "'" + NombreSubir + "',"
-                    + "'" + textBox_GeneroSubir.Text + "',"
-                    + "'" + textBox_PlataformaSubir.Text + "',"
-                    + "'" + textBox_ModalidadSubir.Text + "',"
-                    + "'" + textBox_PrecioSubir.Text + "',"
-                    + "'" + textBox_StockSubir.Text + "')";
-
-                // Cargamos neestra linea de comandos
-                MySqlCommand comando = new MySqlCommand(consulta, conexion);
-
-                // Ejecutamos el comando
-                comando.ExecuteNonQuery();
-
-                MessageBox.Show("Producto agregado correctamente");
-
-                // Cerramos nuestra conexion
-                conexion.Close();
+                MessageBox.Show("No es posible subir el registro (Maximo 10 registros)");
             }
-            catch (Exception ex)
+            else
             {
-                MessageBox.Show("Error al eliminar producto: " + ex.Message);
+                try
+                {
+
+                    // Creamos nuestra variable para la base de datos, y pasamos nuestra informacion
+                    MySqlConnection conexion = new MySqlConnection("Server=localhost; Database=proyecto; User=root; Password=; Sslmode=none;");
+                    // Abrimos nuestra base de datos
+                    conexion.Open();
+
+                    // Linea de comando de SQL
+                    string consulta = "INSERT INTO juegos (ID, Nombre, Imagen, Genero, Plataforma, Modalidad, Precio, Stock) VALUES ("
+                        + "'" + textBox_IDSubir.Text + "',"
+                        + "'" + textBox_NombreSubir.Text + "',"
+                        + "'" + NombreSubir + "',"
+                        + "'" + textBox_GeneroSubir.Text + "',"
+                        + "'" + textBox_PlataformaSubir.Text + "',"
+                        + "'" + textBox_ModalidadSubir.Text + "',"
+                        + "'" + textBox_PrecioSubir.Text + "',"
+                        + "'" + textBox_StockSubir.Text + "')";
+
+                    // Cargamos neestra linea de comandos
+                    MySqlCommand comando = new MySqlCommand(consulta, conexion);
+
+                    // Ejecutamos el comando
+                    comando.ExecuteNonQuery();
+
+                    MessageBox.Show("Producto agregado correctamente");
+
+                    // Cerramos nuestra conexion
+                    conexion.Close();
+                }
+                catch (Exception ex)
+                {
+                    MessageBox.Show("Error al eliminar el registro: " + ex.Message);
+                }
             }
 
             // Volvemos a vaciar nuestros TextBox
@@ -206,7 +233,7 @@ namespace ProyectoFinalV1
                 string consulta = "UPDATE juegos SET ID=" + "'" + textBox_IDModificar.Text + "'" +
                     ",Nombre=" + "'" + textBox_NombreModificar.Text + "'" +
                     ",Imagen=" + "'" + NombreModificar + "'" +
-                    ",Genero=" + "'" + textBox_PrecioModificar.Text + "'" +
+                    ",Genero=" + "'" + textBox_GeneroModificar.Text + "'" +
                     ",Plataforma=" + "'" + textBox_PlataformaModificar.Text + "'" +
                     ",Modalidad=" + "'" + textBox_ModalidadModificar.Text + "'" +
                     ",Precio=" + "'" + textBox_PrecioModificar.Text + "'" +
@@ -228,7 +255,7 @@ namespace ProyectoFinalV1
             }
             catch (Exception ex)
             {
-                MessageBox.Show("Error en la actualizacion: " + ex.Message);
+                MessageBox.Show("Error en la actualizacion del registro: " + ex.Message);
             }
 
             // Volvemos a vaciar nuestros TextBox
@@ -245,20 +272,108 @@ namespace ProyectoFinalV1
 
         }
 
-        private void FormAdmin_Load(object sender, EventArgs e)
-        {
-
-        }
-
-        private void txtTITULO_TextChanged(object sender, EventArgs e)
-        {
-
-        }
-
         private void panel1_MouseDown(object sender, MouseEventArgs e)
         {
             ReleaseCapture();
             SendMessage(this.Handle, 0x112, 0xf012, 0);
+        }
+
+        private int Contar_Registros()
+        {
+            int total = 0;
+            try
+            {
+                // Creamos nuestra variable para la base de datos, y pasamos nuestra informacion
+                MySqlConnection conexion = new MySqlConnection("Server=localhost; Database=proyecto; User=root; Password=; Sslmode=none;");
+                // Abrimos nuestra base de datos
+                conexion.Open();
+
+                string consulta = "SELECT COUNT(*) FROM juegos";
+
+                // Cargamos neestra linea de comandos
+                MySqlCommand comando = new MySqlCommand(consulta, conexion);
+
+                // Ejecutamos el comando y guardamos el resultado
+                total = Convert.ToInt32(comando.ExecuteScalar());
+            }
+            catch (Exception ex)
+            {
+                MessageBox.Show("Error al contar los registros: " + ex.Message);
+            }
+
+            return total;
+        }
+
+        // Boton para recuperar los datos del registro a modificar
+        private void button_BuscarModificar_Click(object sender, EventArgs e)
+        {
+            try
+            {
+
+                // Creamos nuestra variable para la base de datos, y pasamos nuestra informacion
+                MySqlConnection conexion = new MySqlConnection("Server=localhost; Database=proyecto; User=root; Password=; Sslmode=none;");
+                // Abrimos nuestra base de datos
+                conexion.Open();
+
+                // Linea de comando de SQL
+                string consulta = "SELECT * FROM juegos WHERE ID='" + Convert.ToInt32(textBox_IDModificar.Text) + "'";
+                // Cargamos nuestro comando
+                MySqlCommand comando = new MySqlCommand(consulta, conexion);
+
+                // Ejecutamos el comando
+                MySqlDataReader lector = comando.ExecuteReader();
+
+                
+                if (lector.Read())
+                {
+                    // Asignamos los valores recuperados a los textBox correspondientes
+                    textBox_NombreModificar.Text = lector["Nombre"].ToString();
+                    textBox_GeneroModificar.Text = lector["Genero"].ToString();
+                    textBox_PlataformaModificar.Text = lector["Plataforma"].ToString();
+                    textBox_ModalidadModificar.Text = lector["Modalidad"].ToString();
+                    textBox_PrecioModificar.Text = lector["Precio"].ToString();
+                    textBox_StockModificar.Text = lector["Stock"].ToString();
+
+                    // Recuperamos el nombre de la imagen asociada al registro
+                    string nombreImagen = lector["Imagen"].ToString();
+
+                    // Tambien guardamos el nombre de la imagen en nuestra variable declarada anteriormente
+                    NombreModificar = lector["Imagen"].ToString();
+
+                    // Construimos la ruta hacia la carpeta donde estan las imagenes de los juegos
+                    string rutaImagenes = System.IO.Path.Combine(Application.StartupPath, "Juegos");
+
+                    // Construimos la ruta completa donde se encuentra nuestra imagen a buscar
+                    string rutaCompleta = System.IO.Path.Combine(rutaImagenes, nombreImagen);
+
+                    // Si hay un archivo con esta direccion en especifico
+                    if (System.IO.File.Exists(rutaCompleta))
+                    {
+                        // Cargamos la imagen a nuestro picture box
+                        pictureBox_Modificar.Image = Image.FromFile(rutaCompleta);
+                    }
+                    else
+                    {
+                        // En caso de que la imagen no exista, entonces mostramos mensaje de error
+                        MessageBox.Show("No se encontro la imagen deseada");
+
+                        // Si habia una imagen en el pictureBox, la quitamos
+                        pictureBox_Modificar.Image = null; 
+                    }
+
+                }
+                else
+                {
+                    MessageBox.Show("Error al encontrar el registro");
+                }
+
+                lector.Close();
+                conexion.Close();
+            }
+            catch (Exception ex)
+            {
+                MessageBox.Show("Error al buscar el registro: " + ex.Message);
+            }
         }
     }
 }
